Add function to release linked list
A new function 'ReleaseList' has been introduced to handle the proper deletion of the linked list nodes. It traverses the list, deletes each node, and releases the memory back to the system. This function is called in the main function to ensure safe disposal of the list at program termination.