Implement InsertAtTail function in doubly linked list
The commit introduces the InsertAtTail function to the doubly linked list code, to add a node at the end of the list. Changes also include improved print output for easy debugging and minor stylistic adjustments. The operation of adding nodes at the head of the list in the main function has been commented out and replaced with the addition at the tail.