Fix insertAtHead method and revise tests
The insertAtHead method in the SingleLinkedList class was refactored to consistently insert the new node at the head of the list, regardless of the list's current size. An exception was also added for scenarios where the size could potentially be less than zero. Corresponding changes were made to the tests in SingleLinkedListTest to reflect these modifications.